Start work on removing gpgme as a dependency
The library is horrible (particularly in terms of documentation). We're
going to switch over to python-gnupg as soon as we can.

This patch reimplements the hash algorithm -> name function without
depending on gpgme constants. We also collapse the two functions, since
there's no need to derive the non-conformant intermediate result.